+++ src/core_modules/capture-core/components/DataEntries/SingleEventRegistrationEntry/DataEntryWrapper/DataEntry/epics/getConvertedNewSingleEvent.js
@@ -43,6 +43,44 @@ export const getNewEventServerData = (state: ReduxState, formFoundation: RenderF
     };
 };
 
+export const getAddEventEnrollmentServerData = (state: ReduxState,
+    formFoundation: RenderFoundation,
+    formClientValues: Object,
+    mainDataClientValues: Object,
+    completed?: boolean,
+) => {
+    const formServerValues = formFoundation.convertValues(formClientValues, convertToServerValue);
+    const mainDataServerValues: Object = convertMainEventClientToServer(mainDataClientValues);
+    const { teiId, enrollmentId, programId, orgUnitId } = state.router.location.query;
+
+    if (!mainDataServerValues.status) {
+        mainDataServerValues.status = completed ? 'ACTIVE' : 'COMPLETED';
+    }
+    if (mainDataServerValues.status === 'COMPLETED') {
+        mainDataServerValues.completedDate = getFormattedStringFromMomentUsingEuropeanGlyphs(moment());
+    }
+
+    return {
+        events: [
+            {
+                ...mainDataServerValues,
+                program: programId,
+                programStage: formFoundation.id,
+                orgUnit: orgUnitId,
+                trackedEntityInstance: teiId,
+                enrollment: enrollmentId,
+                ...getApiCategoriesArgument(state.currentSelections.categories),
+                dataValues: Object
+                    .keys(formServerValues)
+                    .map(key => ({
+                        dataElement: key,
+                        value: formServerValues[key],
+                    })),
+            },
+        ],
+    };
+};
+
 function getDataEntriesNotes(state: ReduxState, dataEntryKey: string) {
     const notes = state.dataEntriesNotes && state.dataEntriesNotes[dataEntryKey];
     return notes ? notes.map(note => ({ value: note.value })) : [];