+++ src/core_modules/capture-core/components/DataEntries/SingleEventRegistrationEntry/SingleEventRegistrationEntry.container.js
@@ -1,20 +1,13 @@
 // @flow
 import { connect } from 'react-redux';
 import { type ComponentType } from 'react';
-import i18n from '@dhis2/d2-i18n';
 import { compose } from 'redux';
 import { SingleEventRegistrationEntryComponent } from './SingleEventRegistrationEntry.component';
 import { withBrowserBackWarning } from '../../../HOC/withBrowserBackWarning';
 import { dataEntryHasChanges } from '../../DataEntry/common/dataEntryHasChanges';
 import { makeEventAccessSelector } from './SingleEventRegistrationEntry.selectors';
 import { withLoadingIndicator } from '../../../HOC';
-
-const dialogConfig = {
-    header: i18n.t('Unsaved changes'),
-    text: i18n.t('Leaving this page will discard the changes you made to this event.'),
-    confirmText: i18n.t('Yes, discard'),
-    cancelText: i18n.t('No, stay here'),
-};
+import { defaultDialogProps as dialogConfig } from '../../Dialogs/ConfirmDialog.constants';
 
 const inEffect = (state: ReduxState) => dataEntryHasChanges(state, 'singleEvent-newEvent') || state.newEventPage.showAddRelationship;